+++ doc/changelog.rst
@@ -13,6 +13,7 @@ Core
 ~~~~
 
 - Move calculate_taxes_automatically from ``OrderSource`` to ``TaxModule``
+- Add shop product validation for OrderSource
 - Add contact address for ``Shop``
 - Add update_stock calls for ``SimpleSupplierModule``
 - Rename ``simple_pricing`` to ``customer_group_pricing``